refactor(theme): split lib.rs into modules by concern
