+++ crates/theme/src/paint.rs
@@ -1,0 +1,266 @@
+//! Context-free paint helpers: the process-wide appearance mirror and the
+//! free functions ([`ink`], [`hairline`], [`wash`], …) that element builders
+//! call without a `cx` in scope.
+
+use std::sync::atomic::{AtomicU8, AtomicU32, Ordering};
+
+use gpui::{BoxShadow, Hsla, hsla, point, px};
+
+use crate::Appearance;
+
+/// Process-wide mirror of the installed theme's appearance.
+///
+/// The paint helpers ([`ink`], [`hairline`], [`wash`], …) are free functions
+/// called from deep inside element builders that have no `cx` in scope, so they
+/// read the appearance from here instead of the gpui global. Appearance is
+/// genuinely process-wide — one setting for every window — so a single mirror is
+/// sound; [`Theme::install`](crate::theme::Theme::install) is the only writer
+/// outside tests.
+static CURRENT_APPEARANCE: AtomicU8 = AtomicU8::new(0);
+
+/// Bumped every time the appearance actually changes.
+///
+/// Anything that caches *resolved colors* — most importantly the markdown
+/// renderer's cross-frame `TextRun` cache, which bakes an `Hsla` into every run —
+/// is only valid for the palette that produced it. Those caches were written when
+/// the theme was a compile-time constant, so their validity keys cover content
+/// only. Rather than thread the palette through every key, they compare this
+/// counter and drop everything when it moves.
+static THEME_GENERATION: AtomicU32 = AtomicU32::new(0);
+
+/// The appearance the context-free paint helpers are painting for.
+pub fn current_appearance() -> Appearance {
+    match CURRENT_APPEARANCE.load(Ordering::Relaxed) {
+        1 => Appearance::Light,
+        _ => Appearance::Dark,
+    }
+}
+
+/// Monotonic id of the current palette.
+pub fn theme_generation() -> u32 {
+    THEME_GENERATION.load(Ordering::Relaxed)
+}
+
+/// [`CURRENT_APPEARANCE`] is process-wide, so under the parallel test runner
+/// any test that flips it — or asserts on the output of a helper that reads it
+/// ([`ink`], [`hairline`], [`wash`], …) — must hold this lock. Crate-visible
+/// because such tests exist outside this module too (see bezel-motion's tests).
+/// Tests that flip the appearance restore Dark before releasing the guard.
+/// Compiled unconditionally (not `cfg(test)`) so downstream crates' tests can
+/// use it too; it is not part of the public API.
+#[doc(hidden)]
+pub fn lock_appearance() -> std::sync::MutexGuard<'static, ()> {
+    static APPEARANCE_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
+    APPEARANCE_LOCK.lock().unwrap_or_else(|e| e.into_inner())
+}
+
+/// Point the context-free paint helpers at an appearance. Called by
+/// [`Theme::install`](crate::theme::Theme::install); exposed for tests that
+/// build a theme without an `App`.
+pub fn set_current_appearance(appearance: Appearance) {
+    let encoded = match appearance {
+        Appearance::Dark => 0,
+        Appearance::Light => 1,
+    };
+    if CURRENT_APPEARANCE.swap(encoded, Ordering::Relaxed) != encoded {
+        THEME_GENERATION.fetch_add(1, Ordering::Relaxed);
+    }
+}
+
+/// Light-mode alpha multiplier for **fills** (hover/active washes, chip and pill
+/// backgrounds).
+///
+/// This was 0.5 on the theory that dark ink on a bright field reads heavier and
+/// should be scaled back. That theory is right for a *large* wash and badly wrong
+/// for everything else: this palette leans on very low alphas for its subtle
+/// fills — the composer plate is `ink(0.03)`, key caps are `ink(0.05)` — and
+/// halving those produced 1.5% black on white, which is nothing. The composer
+/// lost its background entirely and selected tabs stopped reading as selected.
+///
+/// The established light-UI scales (Primer, Radix) land subtle ≈ 3–4%, hover ≈ 8%,
+/// selected ≈ 14% black — which is where the dark palette's white alphas already
+/// sit. So the honest multiplier is 1: the same number in both appearances, with
+/// only the *tone* flipping. Any per-state correction belongs in that state's
+/// token, not in a blanket multiplier.
+pub const INK_FILL_SCALE: f32 = 1.0;
+
+/// Light-mode alpha multiplier for **hairlines** (borders, dividers, rings).
+/// Opposite of fills: a 1px edge has to hold its own against a bright surround,
+/// and the dark palette's white hairlines are deliberately faint. Scaling up
+/// keeps separators legible instead of dissolving into the panel.
+pub const INK_HAIRLINE_SCALE: f32 = 1.35;
+
+/// Translucent **fill** ink for interactive states and chip plates: soft-white on
+/// dark, soft-black on light at [`INK_FILL_SCALE`] of the alpha.
+///
+/// Alphas are quoted in *dark-mode terms* at every call site — the dark theme is
+/// the tuned one — and the light value is derived. Callers keep one number and
+/// both appearances stay in the relationship the dark tuning established.
+///
+/// Fills must never rest on transparent BLACK in dark mode: fully opaque washes
+/// killed the glass and flashed dark mid-fade (user reports), so hover fades rest
+/// on `ink(0.0)`, which stays tonally correct at zero alpha.
+pub fn ink(alpha: f32) -> Hsla {
+    ink_for(current_appearance(), alpha)
+}
+
+pub(crate) fn ink_for(appearance: Appearance, alpha: f32) -> Hsla {
+    match appearance {
+        // Soft-white, not pure white: alphas are high enough to stay visible at
+        // the brightest backdrop the 0.90 glass scrim can produce.
+        Appearance::Dark => hsla(0.0, 0.0, 1.0, alpha),
+        Appearance::Light => hsla(0.0, 0.0, 0.0, alpha * INK_FILL_SCALE),
+    }
+}
+
+/// Translucent **hairline** ink for borders, dividers and rings: white on dark,
+/// black on light at [`INK_HAIRLINE_SCALE`] of the alpha.
+///
+/// Separate from [`ink`] because edges and fills scale in opposite directions
+/// when the field brightens — a 1px line needs *more* ink on white, a plate needs
+/// less.
+pub fn hairline(alpha: f32) -> Hsla {
+    hairline_for(current_appearance(), alpha)
+}
+
+pub(crate) fn hairline_for(appearance: Appearance, alpha: f32) -> Hsla {
+    match appearance {
+        Appearance::Dark => hsla(0.0, 0.0, 1.0, alpha),
+        Appearance::Light => hsla(0.0, 0.0, 0.0, (alpha * INK_HAIRLINE_SCALE).min(0.5)),
+    }
+}
+
+/// Interactive-state wash: a softened [`ink`] that stops short of pure black or
+/// white so hover plates read as tinted glass rather than paint.
+pub fn wash(alpha: f32) -> Hsla {
+    wash_for(current_appearance(), alpha)
+}
+
+pub(crate) fn wash_for(appearance: Appearance, alpha: f32) -> Hsla {
+    match appearance {
+        Appearance::Dark => hsla(0.0, 0.0, 0.92, alpha),
+        Appearance::Light => hsla(0.0, 0.0, 0.10, alpha * INK_FILL_SCALE),
+    }
+}
+
+/// Alpha of the standard modal backdrop in dark mode. Call sites that need a
+/// heavier or lighter scrim pass their own dark-mode alpha to [`scrim`].
+pub const SCRIM_ALPHA_DARK: f32 = 0.60;
+
+/// Modal backdrop at `alpha_dark` (quoted, as everywhere, in dark-mode terms).
+///
+/// Black in both appearances — a scrim's job is to darken what is behind it, and
+/// a "light scrim" of white would wash the modal out rather than seat it. What
+/// changes is strength: on a bright field a dark-mode-weight scrim reads as a
+/// blackout, so light mode scales to roughly half.
+pub fn scrim(alpha_dark: f32) -> Hsla {
+    scrim_for(current_appearance(), alpha_dark)
+}
+
+pub(crate) fn scrim_for(appearance: Appearance, alpha_dark: f32) -> Hsla {
+    match appearance {
+        Appearance::Dark => hsla(0.0, 0.0, 0.0, alpha_dark),
+        Appearance::Light => hsla(0.0, 0.0, 0.0, 0.32 * (alpha_dark / SCRIM_ALPHA_DARK)),
+    }
+}
+
+/// Recessed band behind a palette/picker header or footer strip.
+///
+/// A free function as well as a [`Theme`](crate::theme::Theme) field because the
+/// picker chrome that paints it is built from context-free helpers; both resolve
+/// to the same value.
+pub fn band() -> Hsla {
+    band_for(current_appearance())
+}
+
+pub(crate) fn band_for(appearance: Appearance) -> Hsla {
+    match appearance {
+        Appearance::Dark => hsla(0.0, 0.0, 0.0, 0.16),
+        // A recessed strip on white needs far less ink than on near-black; the
+        // dark 16% would read as a bruise.
+        Appearance::Light => hsla(0.0, 0.0, 0.0, 0.045),
+    }
+}
+
+/// Selected-state glass treatment (tabs, session rows, space rows): a
+/// TRANSLUCENT wash the vibrancy reads through — heavier flat washes blocked
+/// the glass (user request). Dark: the 11% [`wash`]. Light: the tone-flipped
+/// wash at 6% — 11% black read too dark over the bright frost (user report;
+/// light also previously ran a near-opaque white chip, rejected the same
+/// way). Same fill as [`Theme::glass_hover`](crate::theme::Theme::glass_hover) —
+/// the ring in [`glass_selected_shadows`] is what distinguishes selection.
+/// Selection *inside floating cards* is different — see [`card_selected_bg`].
+pub fn glass_selected_bg() -> Hsla {
+    match current_appearance() {
+        Appearance::Dark => wash(0.11),
+        Appearance::Light => wash(0.06),
+    }
+}
+
+/// The user message bubble's plate: the same translucent wash family as
+/// [`glass_selected_bg`], one step softer — at the selection weight the
+/// bubble read too strong for settled content (user report), and an opaque
+/// plate before that read as a solid slab over glass.
+pub fn user_bubble_bg() -> Hsla {
+    match current_appearance() {
+        Appearance::Dark => wash(0.08),
+        Appearance::Light => wash(0.04),
+    }
+}
+
+/// Selected/keyboard-active treatment for rows and chips INSIDE a floating
+/// card (menu rows, the picker rail, segmented chips). The card is already the
+/// bright plane in light mode, so a white lift can't read there — selection is
+/// the tone-flipped grey wash, at 6% (dark's 11% read too dark on the bright
+/// plane, user report).
+pub fn card_selected_bg() -> Hsla {
+    match current_appearance() {
+        Appearance::Dark => wash(0.11),
+        Appearance::Light => wash(0.06),
+    }
+}
+
+/// The selected chip's bright outline, as an INSET shadow: gpui paints inset
+/// shadows ON TOP of the background, edges only — a border with zero layout
+/// cost. Drop shadows are filled rects painted BEHIND the element, and behind
+/// a 5% fill they showed straight through as an opaque dark plate with a
+/// greyed ring (user report) — nothing may paint behind a glass chip.
+///
+/// Light pins the ring at a flat 7% black rather than the scaled hairline:
+/// heavier rings (the [`INK_HAIRLINE_SCALE`]d value, then 12%) outlined every
+/// selected chip in a dark box (user reports) — the ring should define the
+/// chip the way dark's 9% white ring does, not frame it.
+///
+/// There is deliberately NO drop-shadow seat under the light chip. Three
+/// recipes were tried (a tight 10% layer, a 6% contact + 5% ambient pair, a
+/// lone 4% whisper) and every one failed on sight: layers sum into a grey rim
+/// exactly where the chip meets the frost, gpui's small-radius blur reads
+/// coarse on a bright field, and the tab strip is a scroll container that
+/// clips its children vertically — any shadow escaping the chip gets cut off
+/// mid-fade. The near-opaque fill plus the ring carry selection, exactly as
+/// dark's wash plus ring does; the two appearances share one recipe now.
+pub fn glass_selected_shadows() -> Vec<BoxShadow> {
+    card_selected_shadows()
+}
+
+/// Selection outline for rows and chips INSIDE a floating card (menu rows,
+/// the picker rail, segmented chips): the inset ring alone, in both
+/// appearances. Card rows fill with a translucent wash
+/// ([`card_selected_bg`]), and a drop shadow — a filled rect painted BEHIND
+/// the element — shows straight through a translucent fill as a grey plate
+/// (the same lesson [`glass_selected_shadows`] records for dark glass). The
+/// card already carries the elevation shadow; selection inside it only needs
+/// the edge.
+pub fn card_selected_shadows() -> Vec<BoxShadow> {
+    let color = match current_appearance() {
+        Appearance::Dark => hairline(0.09),
+        Appearance::Light => hsla(0.0, 0.0, 0.0, 0.07),
+    };
+    vec![BoxShadow {
+        color,
+        offset: point(px(0.0), px(0.0)),
+        blur_radius: px(0.0),
+        spread_radius: px(1.0),
+        inset: true,
+    }]
+}
